removed serverside libs from appster to add them within bundler and appster-server
