+++ packages/@coorpacademy-appster/README.md
@@ -44,44 +44,3 @@ import createRoutes from './routes';
 
 const app = prepareApp(createReducer, createRoutes, options);
 ```
-
-## webpack config use
-```
-const path = require('path');
-const {webpackConfig} = require('@coorpacademy/appster');
-
-const input = path.join(__dirname, 'src/app');
-const output = path.join(__dirname, 'dist');
-
-module.exports = webpackConfig('Forum', input, output);
-```
-
-## sandbox app use
-
-Give reducers and routes as params
-```js
-const path = require('path');
-const config = require('./config');
-const {runServer} = require('@coorpacademy/appster');
-const input = path.join(__dirname, '../app');
-
-const addSandbox = app => {
-  return app.get('/', (req, res) => {
-    res.send(`
-      <div id="yourapp"></div>
-      <script type="text/javascript" src="/dist/your-app.js"></script>
-      <script>
-        var options = {
-          container: document.getElementById('yourapp'),
-          foo: 'bar',
-          api: 'http://localhost:3000/api/v1'
-        };
-
-        YourApp.create(options);
-      </script>
-    `);
-  });
-};
-
-runServer('YourApp', config, input, addSandbox);
-```
